[Spark] Share DeltaGeoSuite between 4.1 and 4.2 via a base trait
- Extract the 55 tests identical across Spark 4.1 and 4.2 into a shared `DeltaGeoSuiteBase` trait in `scala-shims/spark-4.1-4.2`.
- Reduce each per-version `DeltaGeoSuite` to a thin subclass holding only its unique tests:
  - 4.1: 1 test (CONVERT TO DELTA at the validation-hook level).
  - 4.2: 16 tests (INSERT/UPDATE/DELETE, nested/array/map geo, MERGE SCHEMA EVOLUTION, CDF lifecycle, end-to-end CONVERT) that rely on 4.2's Parquet writer encoding geo types.
- Removes ~700 lines of duplicated test code (3 files, +791 / -1498).

- The two suites shared 55 byte-identical tests; only the DML/CDF paths genuinely diverge (OSS Spark 4.1's Parquet writer cannot encode `GeometryType`/`GeographyType`).

- `DeltaGeoSuite` runs green with unchanged counts: 56 tests on 4.1, 71 on 4.2.
