+++ spark/src/test/scala-shims/spark-4.2/org/apache/spark/sql/delta/DeltaGeoSuite.scala
@@ -16,43 +16,19 @@
 
 package org.apache.spark.sql.delta
 
-import org.apache.spark.sql.delta.actions.{Metadata, Protocol}
-import org.apache.spark.sql.delta.catalog.DeltaTableV2
-import org.apache.spark.sql.delta.schema.SchemaMergingUtils
-import org.apache.spark.sql.delta.shims.GeoTypesShim
-import org.apache.spark.sql.delta.sources.DeltaSQLConf
-import org.apache.spark.sql.delta.test.{DeltaSQLCommandTest, DeltaSQLTestUtils}
-import org.apache.spark.sql.delta.test.DeltaTestImplicits._
-
-import org.apache.spark.{SparkConf, SparkThrowable}
-import org.apache.spark.sql.{AnalysisException, QueryTest}
+import org.apache.spark.sql.AnalysisException
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.expressions.st.{
-  ST_AsBinary, ST_GeogFromWKB, ST_GeomFromWKB, ST_SetSrid, ST_Srid}
-import org.apache.spark.sql.test.SharedSparkSession
 import org.apache.spark.sql.types._
 
 /**
- * Tests for the GeoSpatial table feature in the Delta Spark connector. Verifies:
- *  - Auto-enablement of the [[GeoSpatialTableFeature]] when a geospatial column is present.
- *  - Required protocol versions (3, 7).
- *  - SRID validation at commit time.
- *  - Removability contract (drop is rejected while geospatial columns still exist).
+ * Spark 4.2 GeoSpatial suite. The shared assertions live in [[DeltaGeoSuiteBase]]
+ * (`scala-shims/spark-4.1-4.2`); this subclass adds the 4.2-only coverage.
  *
- * Note: this suite lives under the spark-4.1 test shim directory because Spark's
- * `GeometryType` / `GeographyType` catalyst types were only added in Spark 4.1
- * (SPARK-53760). The GeoSpatial table feature is therefore a no-op on Spark 4.0.
+ * Spark 4.2's Parquet writer can encode `GeometryType`/`GeographyType`, so this suite adds
+ * the end-to-end read/write CUJ (INSERT/UPDATE/DELETE, nested/array/map geo, MERGE WITH SCHEMA
+ * EVOLUTION, CDF lifecycle, and CONVERT TO DELTA) that OSS Spark 4.1 cannot run.
  */
-class DeltaGeoSuite extends QueryTest
-  with SharedSparkSession
-  with DeltaSQLCommandTest
-  with DeltaSQLTestUtils {
-
-  override protected def sparkConf: SparkConf =
-    super.sparkConf.set(DeltaSQLConf.DELTA_GEO_PREVIEW_ENABLED.key, "true")
-
-  // Default SRID used in Parquet/Iceberg/Delta specs (OGC:CRS84).
-  private val DefaultSrid = 4326
+class DeltaGeoSuite extends DeltaGeoSuiteBase {
 
   // Pre-computed little-endian WKB hex literals for a few well-known points. WKB layout:
   //   byteOrder (1 byte, 0x01 = little-endian)
@@ -68,582 +44,6 @@ class DeltaGeoSuite extends QueryTest
   /** Convert a byte array to an uppercase hex string for direct comparison with WKB literals. */
   private def bytesToHex(bytes: Array[Byte]): String =
     bytes.map(b => f"${b & 0xff}%02X").mkString
-
-  private def metadataWithSchema(schema: StructType): Metadata = {
-    Metadata(schemaString = schema.json)
-  }
-
-  test("geo preview is disabled by default") {
-    assert(DeltaSQLConf.DELTA_GEO_PREVIEW_ENABLED.defaultValue.contains(false),
-      "DELTA_GEO_PREVIEW_ENABLED must default to false while geospatial is in private preview")
-  }
-
-  test("containsGeoColumns detects top-level geometry and geography columns") {
-    val schemaGeom = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeometryType(DefaultSrid))
-    val schemaGeog = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeographyType(DefaultSrid))
-    val schemaPlain = new StructType()
-      .add("id", IntegerType)
-      .add("s", StringType)
-
-    assert(DeltaGeoSpatial.containsGeoColumns(schemaGeom))
-    assert(DeltaGeoSpatial.containsGeoColumns(schemaGeog))
-    assert(!DeltaGeoSpatial.containsGeoColumns(schemaPlain))
-  }
-
-  test("containsGeoColumns detects nested geospatial columns") {
-    val nested = new StructType()
-      .add("outer", new StructType().add("inner", GeometryType(DefaultSrid)))
-    assert(DeltaGeoSpatial.containsGeoColumns(nested))
-
-    val arr = new StructType().add("xs", ArrayType(GeometryType(DefaultSrid)))
-    assert(DeltaGeoSpatial.containsGeoColumns(arr))
-
-    val map = new StructType().add("m", MapType(StringType, GeographyType(DefaultSrid)))
-    assert(DeltaGeoSpatial.containsGeoColumns(map))
-  }
-
-  test("findGeoColumnsRecursively returns the first matching type") {
-    val schema = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeographyType(DefaultSrid))
-    val result = DeltaGeoSpatial.findGeoColumnsRecursively(schema)
-    assert(result.isDefined)
-    assert(result.get.isInstanceOf[GeographyType])
-  }
-
-  test("assertSridSupported passes for default fixed SRIDs") {
-    val schema = new StructType()
-      .add("g1", GeometryType(DefaultSrid))
-      .add("g2", GeographyType(DefaultSrid))
-    DeltaGeoSpatial.assertSridSupported(schema)
-  }
-
-  test("isGeoSpatialType returns true for geometry and geography") {
-    assert(DeltaGeoSpatial.isGeoSpatialType(GeometryType(DefaultSrid)))
-    assert(DeltaGeoSpatial.isGeoSpatialType(GeographyType(DefaultSrid)))
-    assert(!DeltaGeoSpatial.isGeoSpatialType(StringType))
-  }
-
-  test("DeltaDataSource accepts GeoSpatial types via supportsDataType") {
-    val ds = new sources.DeltaDataSource
-    assert(ds.supportsDataType(GeometryType(DefaultSrid)))
-    assert(ds.supportsDataType(GeographyType(DefaultSrid)))
-  }
-
-  test("isSupported checks both preview and stable features") {
-    val protoNone = Protocol(3, 7)
-    val protoStable = Protocol(3, 7).withFeature(GeoSpatialTableFeature)
-    val protoPreview = Protocol(3, 7).withFeature(GeoSpatialPreviewTableFeature)
-    assert(!DeltaGeoSpatial.isSupported(protoNone))
-    assert(DeltaGeoSpatial.isSupported(protoStable))
-    assert(DeltaGeoSpatial.isSupported(protoPreview))
-  }
-
-  test("GeoSpatialTableFeature is a reader+writer feature with min protocol (3, 7)") {
-    assert(GeoSpatialTableFeature.isReaderWriterFeature)
-    assert(GeoSpatialTableFeature.minReaderVersion == 3)
-    assert(GeoSpatialTableFeature.minWriterVersion == 7)
-    assert(GeoSpatialTableFeature.name == "geospatial")
-    assert(GeoSpatialPreviewTableFeature.name == "geospatial-dev")
-  }
-
-  test("Both GeoSpatial features are registered in allSupportedFeaturesMap") {
-    assert(TableFeature.featureNameToFeature("geospatial").contains(GeoSpatialTableFeature))
-    assert(TableFeature.featureNameToFeature("geospatial-dev")
-      .contains(GeoSpatialPreviewTableFeature))
-  }
-
-  test("metadataRequiresFeatureToBeEnabled is true iff schema has geospatial columns") {
-    val withGeo = metadataWithSchema(new StructType().add("g", GeometryType(DefaultSrid)))
-    val withoutGeo = metadataWithSchema(new StructType().add("s", StringType))
-    assert(GeoSpatialTableFeature.metadataRequiresFeatureToBeEnabled(
-      Protocol(1, 2), withGeo, spark))
-    assert(!GeoSpatialTableFeature.metadataRequiresFeatureToBeEnabled(
-      Protocol(1, 2), withoutGeo, spark))
-  }
-
-  test("stable feature does not auto-enable when preview feature is already supported") {
-    val withGeo = metadataWithSchema(new StructType().add("g", GeometryType(DefaultSrid)))
-    val protocolWithPreview =
-      Protocol(3, 7).withFeature(GeoSpatialPreviewTableFeature)
-    assert(!GeoSpatialTableFeature.metadataRequiresFeatureToBeEnabled(
-      protocolWithPreview, withGeo, spark))
-  }
-
-  test("automaticallyUpdateProtocolOfExistingTables returns true") {
-    assert(GeoSpatialTableFeature.automaticallyUpdateProtocolOfExistingTables)
-  }
-
-  test("creating a table with a geometry column auto-enables GeoSpatial feature") {
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta")
-      val protocol = getProtocolForTable("tbl")
-      assert(protocol.isFeatureSupported(GeoSpatialTableFeature),
-        s"Expected GeoSpatialTableFeature in protocol but got: $protocol")
-      assert(protocol.minReaderVersion >= GeoSpatialTableFeature.minReaderVersion)
-      assert(protocol.minWriterVersion >= GeoSpatialTableFeature.minWriterVersion)
-    }
-  }
-
-  test("creating a table without geospatial columns does not enable GeoSpatial feature") {
-    withTable("tbl") {
-      sql("CREATE TABLE tbl(id INT, s STRING) USING delta")
-      val protocol = getProtocolForTable("tbl")
-      assert(!protocol.isFeatureSupported(GeoSpatialTableFeature))
-      assert(!protocol.isFeatureSupported(GeoSpatialPreviewTableFeature))
-    }
-  }
-
-  test("ALTER TABLE ADD COLUMN with geography auto-enables GeoSpatial feature") {
-    withTable("tbl") {
-      sql("CREATE TABLE tbl(id INT) USING delta")
-      assert(!getProtocolForTable("tbl").isFeatureSupported(GeoSpatialTableFeature))
-      sql(s"ALTER TABLE tbl ADD COLUMN g GEOGRAPHY($DefaultSrid)")
-      assert(getProtocolForTable("tbl").isFeatureSupported(GeoSpatialTableFeature))
-    }
-  }
-
-  test("PartitionUtils rejects GeoSpatial type as a partition column") {
-    // Geometry / Geography are not orderable and thus cannot be partition columns. Delta's
-    // `util/PartitionUtils.validatePartitionColumn` is the write-path backstop; the catalog
-    // (DDL) path goes through Spark's own `PartitioningUtils` which is out of Delta's hands.
-    val schema = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeometryType(DefaultSrid))
-    val ex = intercept[DeltaAnalysisException] {
-      util.PartitionUtils.validatePartitionColumn(schema, Seq("g"), caseSensitive = false)
-    }
-    assert(ex.getErrorClass == "DELTA_INVALID_PARTITION_COLUMN_TYPE",
-      s"Unexpected error class: ${ex.getErrorClass}")
-  }
-
-  test("commit fails when geo column is added and preview conf is disabled") {
-    withTable("tbl") {
-      sql("CREATE TABLE tbl(id INT) USING delta")
-      withSQLConf(DeltaSQLConf.DELTA_GEO_PREVIEW_ENABLED.key -> "false") {
-        val ex = intercept[SparkThrowable] {
-          sql(s"ALTER TABLE tbl ADD COLUMN g GEOMETRY($DefaultSrid)")
-        }
-        assert(ex.getErrorClass == "DELTA_GEOSPATIAL_NOT_SUPPORTED",
-          s"Unexpected error class: ${ex.getErrorClass}")
-      }
-    }
-  }
-
-  test("read fails when table has geo column and preview conf is disabled") {
-    // Covers the read-path guard wired through [[ProtocolMetadataAdapterV1.assertTableReadable]]:
-    // a `SELECT *` against a table with a geo column triggers DeltaParquetFileFormat
-    // construction, which invokes [[DeltaGeoSpatial.assertTableReadable]] and throws when the
-    // preview config is off.
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta")
-      withSQLConf(DeltaSQLConf.DELTA_GEO_PREVIEW_ENABLED.key -> "false") {
-        val ex = intercept[SparkThrowable] {
-          sql("SELECT * FROM tbl").collect()
-        }
-        assert(ex.getErrorClass == "DELTA_GEOSPATIAL_NOT_SUPPORTED",
-          s"Unexpected error class: ${ex.getErrorClass}")
-      }
-    }
-  }
-
-  test("assertMetadata refuses a geo column in a committed Metadata action when preview is off") {
-    // Covers the inline guard in [[OptimisticTransactionImpl.assertMetadata]].
-    // The standard DDL path hits [[GeoSpatialTableFeature.metadataRequiresFeatureToBeEnabled]]
-    // first - see the test above - so we exercise the inline guard by committing a Metadata
-    // action directly, which routes through `prepareCommit -> updateMetadataAndProtocolWith-
-    // RequiredFeatures -> assertMetadata` before the table-feature gate runs.
-    withTable("tbl") {
-      sql("CREATE TABLE tbl(id INT) USING delta")
-      withSQLConf(DeltaSQLConf.DELTA_GEO_PREVIEW_ENABLED.key -> "false") {
-        val log = DeltaLog.forTable(spark, TableIdentifier("tbl"))
-        val txn = log.startTransaction()
-        val geoSchema = new StructType()
-          .add("id", IntegerType)
-          .add("g", GeometryType(DefaultSrid))
-        val geoMetadata = txn.snapshot.metadata.copy(schemaString = geoSchema.json)
-
-        val ex = intercept[SparkThrowable] {
-          txn.commit(Seq(geoMetadata), DeltaOperations.ManualUpdate)
-        }
-        assert(ex.getErrorClass == "UNSUPPORTED_DATATYPE",
-          s"Unexpected error class: ${ex.getErrorClass}")
-      }
-    }
-  }
-
-  test("DROP FEATURE is rejected while geospatial columns remain in the schema") {
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta")
-      assert(getProtocolForTable("tbl").isFeatureSupported(GeoSpatialTableFeature))
-
-      val ex = intercept[DeltaAnalysisException] {
-        sql(s"ALTER TABLE tbl DROP FEATURE ${GeoSpatialTableFeature.name}")
-      }
-      assert(ex.getErrorClass == "DELTA_CANNOT_DROP_GEOSPATIAL_FEATURE",
-        s"Unexpected error class: ${ex.getErrorClass}")
-      assert(ex.getMessage.contains("g"))
-    }
-  }
-
-  test("validateDropInvariants returns true after geospatial columns are dropped") {
-    withTable("tbl") {
-      sql(
-        s"""CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta
-           |TBLPROPERTIES('delta.columnMapping.mode' = 'name')""".stripMargin)
-      val tableBefore = DeltaTableV2(spark, TableIdentifier("tbl"))
-      assert(!GeoSpatialTableFeature.validateDropInvariants(
-        tableBefore, tableBefore.initialSnapshot))
-
-      sql("ALTER TABLE tbl DROP COLUMN g")
-      val tableAfter = DeltaTableV2(spark, TableIdentifier("tbl"))
-      assert(GeoSpatialTableFeature.validateDropInvariants(
-        tableAfter, tableAfter.initialSnapshot))
-    }
-  }
-
-  test("validateDropInvariants flips to true after dropping a struct containing nested geo") {
-    // `validateDropInvariants` checks `containsGeoColumns` recursively, so a geo column
-    // hidden inside a struct should block feature removal exactly like a top-level geo
-    // column does. After the entire struct column is dropped, the table feature must
-    // become droppable again. This protects against a regression where the drop-check
-    // is reduced to a flat field scan instead of a recursive one.
-    withTable("tbl") {
-      sql(
-        s"""CREATE TABLE tbl(id INT, outer STRUCT<inner: GEOMETRY($DefaultSrid)>)
-           |USING delta
-           |TBLPROPERTIES('delta.columnMapping.mode' = 'name')""".stripMargin)
-      val tableBefore = DeltaTableV2(spark, TableIdentifier("tbl"))
-      assert(!GeoSpatialTableFeature.validateDropInvariants(
-        tableBefore, tableBefore.initialSnapshot),
-        "Nested geo inside a struct should block feature removal")
-
-      // Also confirm the user-facing DROP FEATURE rejection fires for nested geo.
-      val ex = intercept[DeltaAnalysisException] {
-        sql(s"ALTER TABLE tbl DROP FEATURE ${GeoSpatialTableFeature.name}")
-      }
-      assert(ex.getErrorClass == "DELTA_CANNOT_DROP_GEOSPATIAL_FEATURE",
-        s"Unexpected error class: ${ex.getErrorClass}")
-
-      sql("ALTER TABLE tbl DROP COLUMN outer")
-      val tableAfter = DeltaTableV2(spark, TableIdentifier("tbl"))
-      assert(GeoSpatialTableFeature.validateDropInvariants(
-        tableAfter, tableAfter.initialSnapshot),
-        "After dropping the struct column, feature removal should be allowed")
-    }
-  }
-
-  test("actionUsesFeature returns false for all actions (history protection disabled)") {
-    val geoMeta = metadataWithSchema(new StructType().add("g", GeometryType(DefaultSrid)))
-    val plainMeta = metadataWithSchema(new StructType().add("s", StringType))
-    assert(!GeoSpatialTableFeature.actionUsesFeature(geoMeta))
-    assert(!GeoSpatialTableFeature.actionUsesFeature(plainMeta))
-    assert(!GeoSpatialPreviewTableFeature.actionUsesFeature(geoMeta))
-  }
-
-  // ---------------------------------------------------------------------------
-  // Geo schema type handling (Delta log JSON ser/de support)
-  // ---------------------------------------------------------------------------
-
-  test("Schema JSON round-trips a top-level geometry column") {
-    val original = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeometryType(DefaultSrid))
-    val roundtripped = DataType.fromJson(original.json).asInstanceOf[StructType]
-    assert(roundtripped === original)
-    assert(roundtripped("g").dataType.isInstanceOf[GeometryType])
-    assert(roundtripped("g").dataType.asInstanceOf[GeometryType].srid == DefaultSrid)
-  }
-
-  test("Schema JSON round-trips a top-level geography column") {
-    val original = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeographyType(DefaultSrid))
-    val roundtripped = DataType.fromJson(original.json).asInstanceOf[StructType]
-    assert(roundtripped === original)
-    assert(roundtripped("g").dataType.isInstanceOf[GeographyType])
-    assert(roundtripped("g").dataType.asInstanceOf[GeographyType].srid == DefaultSrid)
-  }
-
-  test("Schema JSON round-trips nested geo columns inside structs, arrays, and maps") {
-    val nested = new StructType()
-      .add("outer", new StructType().add("inner", GeometryType(DefaultSrid)))
-      .add("arr", ArrayType(GeographyType(DefaultSrid)))
-      .add("map", MapType(StringType, GeometryType(DefaultSrid)))
-    val roundtripped = DataType.fromJson(nested.json).asInstanceOf[StructType]
-    assert(roundtripped === nested)
-  }
-
-  test("assertSridSupported throws AnalysisException for unsupported SRID on Geometry") {
-    // `GeometryType("ANY")` produces a type whose `srid == MIXED_SRID (-1)`. Spark accepts that
-    // type at construction, but `GeometryType.isSridSupported(-1)` is false, so it's exactly the
-    // gap that `DeltaGeoSpatial.assertSridSupported` is meant to catch. (Direct `GeometryType(-1)`
-    // would now throw `ST_INVALID_SRID_VALUE` from the int constructor and never reach Delta.)
-    val schema = new StructType().add("g", GeometryType("ANY"))
-    val ex = intercept[AnalysisException] {
-      DeltaGeoSpatial.assertSridSupported(schema)
-    }
-    assert(ex.getCondition == "DELTA_GEOSPATIAL_SRID_NOT_SUPPORTED")
-  }
-
-  test("assertSridSupported throws AnalysisException for unsupported SRID on Geography") {
-    val schema = new StructType().add("g", GeographyType("ANY"))
-    val ex = intercept[AnalysisException] {
-      DeltaGeoSpatial.assertSridSupported(schema)
-    }
-    assert(ex.getCondition == "DELTA_GEOSPATIAL_SRID_NOT_SUPPORTED")
-  }
-
-  test("assertSridSupported is a no-op for the default supported SRID") {
-    val schema = new StructType()
-      .add("a", GeometryType(DefaultSrid))
-      .add("b", GeographyType(DefaultSrid))
-    // Should not throw.
-    DeltaGeoSpatial.assertSridSupported(schema)
-  }
-
-  test("validateCommitActions rejects unsupported-SRID metadata even when preview is enabled") {
-    val schema = new StructType().add("g", GeometryType("ANY"))
-    val metadata = metadataWithSchema(schema)
-    val protocol = Protocol(
-      GeoSpatialTableFeature.minReaderVersion,
-      GeoSpatialTableFeature.minWriterVersion)
-      .merge(Protocol.forTableFeature(GeoSpatialTableFeature))
-    val ex = intercept[AnalysisException] {
-      DeltaGeoSpatial.validateCommitActions(spark, protocol, Seq(metadata))
-    }
-    assert(ex.getCondition == "DELTA_GEOSPATIAL_SRID_NOT_SUPPORTED")
-  }
-
-  // Negative-SRID regression pins. Delta's `assertSridSupported` delegates the
-  // `isSridSupported` check to Spark catalyst (`GeometryType.isSridSupported` /
-  // `GeographyType.isSridSupported`). These tests document the Spark-side contract Delta
-  // relies on, so a future Spark change that accepts negative SRIDs cannot silently flip
-  // Delta to accepting them too. Negative SRIDs are how the `"ANY"` mixed-SRID type
-  // materialises (srid == MIXED_SRID == -1), which is the only path through which a
-  // negative SRID can reach Delta - `GeometryType(-1)` is rejected by the catalyst int
-  // constructor before any Delta code runs.
-  test("Spark contract: negative SRIDs are not supported for Geometry or Geography") {
-    assert(!GeometryType.isSridSupported(-1),
-      "Spark contract changed: GeometryType.isSridSupported(-1) now returns true. " +
-        "Delta's negative-SRID rejection relies on this returning false.")
-    assert(!GeographyType.isSridSupported(-1),
-      "Spark contract changed: GeographyType.isSridSupported(-1) now returns true. " +
-        "Delta's negative-SRID rejection relies on this returning false.")
-  }
-
-  test("GeometryType(\"ANY\") materialises with srid == -1 (mixed SRID sentinel)") {
-    // Pin the construction path: assertSridSupported relies on `GeometryType("ANY")`
-    // producing a negative `srid` so the unsupported-SRID check below can fire. If this
-    // ever changes, the assertSridSupported tests above would silently start passing for
-    // a no-op reason.
-    val geometryAnySrid = GeometryType("ANY").srid
-    val geographyAnySrid = GeographyType("ANY").srid
-    assert(geometryAnySrid == -1,
-      s"""Expected GeometryType("ANY").srid == -1, got $geometryAnySrid""")
-    assert(geographyAnySrid == -1,
-      s"""Expected GeographyType("ANY").srid == -1, got $geographyAnySrid""")
-  }
-
-  test("assertSridSupported rejects nested geo columns with negative SRIDs") {
-    // Defence-in-depth: a top-level scalar with a bad SRID is already tested; ensure
-    // recursion through `typeExistsRecursively` also catches a bad SRID buried inside a
-    // struct, since that's how the worst-case bug would manifest (a hidden negative SRID
-    // sneaking past the JSON ser/de).
-    val schema = new StructType()
-      .add("outer", new StructType().add("inner", GeometryType("ANY")))
-    val ex = intercept[AnalysisException] {
-      DeltaGeoSpatial.assertSridSupported(schema)
-    }
-    assert(ex.getCondition == "DELTA_GEOSPATIAL_SRID_NOT_SUPPORTED")
-  }
-
-  // ---------------------------------------------------------------------------
-  // DML command support and validation for geospatial columns
-  // ---------------------------------------------------------------------------
-
-  test("GeoTypesShim.geoExpressions contains the expected ST_ catalyst classes") {
-    val expected = Set[Class[_]](
-      classOf[ST_AsBinary],
-      classOf[ST_GeogFromWKB],
-      classOf[ST_GeomFromWKB],
-      classOf[ST_SetSrid],
-      classOf[ST_Srid])
-    assert(GeoTypesShim.geoExpressions == expected)
-  }
-
-  test("AllowedUserProvidedExpressions whitelists the ST_ classes via the shim") {
-    // The geo ST_* classes are appended to `checkConstraintExpressions` (not `expressions`),
-    // so they are usable in CHECK constraints only and not in generated columns. The shim
-    // contributes the five classes on top of the static `checkConstraintExpressions` Set.
-    val whitelist = AllowedUserProvidedExpressions.checkConstraintExpressions
-    assert(whitelist.contains(classOf[ST_AsBinary]))
-    assert(whitelist.contains(classOf[ST_GeogFromWKB]))
-    assert(whitelist.contains(classOf[ST_GeomFromWKB]))
-    assert(whitelist.contains(classOf[ST_SetSrid]))
-    assert(whitelist.contains(classOf[ST_Srid]))
-  }
-
-  test("failIfSchemaHasGeoColumn rejects schemas containing geometry") {
-    val schema = new StructType()
-      .add("id", IntegerType)
-      .add("g", GeometryType(DefaultSrid))
-    val ex = intercept[Throwable] {
-      DeltaGeoSpatial.failIfSchemaHasGeoColumn(schema, "TEST OP")
-    }
-    assert(ex.getMessage.contains("TEST OP"))
-  }
-
-  test("failIfSchemaHasGeoColumn rejects nested geography columns") {
-    val schema = new StructType()
-      .add("outer", new StructType().add("inner", GeographyType(DefaultSrid)))
-    val ex = intercept[Throwable] {
-      DeltaGeoSpatial.failIfSchemaHasGeoColumn(schema, "TEST NESTED")
-    }
-    assert(ex.getMessage.contains("TEST NESTED"))
-  }
-
-  test("failIfSchemaHasGeoColumn is a no-op for non-geo schemas") {
-    val schema = new StructType()
-      .add("id", IntegerType)
-      .add("s", StringType)
-      .add("xs", ArrayType(IntegerType))
-    // Should not throw.
-    DeltaGeoSpatial.failIfSchemaHasGeoColumn(schema, "TEST")
-  }
-
-  test("ZORDER BY a geometry column is rejected with DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES") {
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta")
-      // No INSERT - `validateZorderByColumns` runs against the snapshot schema before any data
-      // is touched. Inserting geo values would fail in OSS Spark's Parquet writer, which doesn't
-      // (yet) know how to write GeometryType.
-      val ex = intercept[AnalysisException] {
-        sql("OPTIMIZE tbl ZORDER BY (g)")
-      }
-      assert(ex.getCondition == "DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES",
-        s"Expected DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES, got: ${ex.getCondition}")
-    }
-  }
-
-  test("Column default values are rejected on geometry columns at CREATE TABLE") {
-    // `CREATE TABLE ... DEFAULT ...` goes through `CreateTable`, which is one of the ops
-    // `OptimisticTransactionImpl.checkColumnDefaults` matches on. (Note: `ALTER TABLE ADD COLUMN
-    // ... DEFAULT ...` is always rejected by Spark with
-    // `WRONG_COLUMN_DEFAULTS_FOR_DELTA_ALTER_TABLE_ADD_COLUMN_NOT_SUPPORTED` regardless of type
-    // and never reaches the Delta-side check.)
-    //
-    // Use a unique table name + explicit LOCATION inside a temp dir: the CREATE TABLE is
-    // expected to throw, but the catalog/Hadoop FS will already have created the table
-    // directory by the time `checkColumnDefaults` fires, and the default `withTable` cleanup
-    // (just `DROP TABLE IF EXISTS`) would leave that directory behind because the table was
-    // never actually registered. Pointing LOCATION at a `withTempDir` path ensures it is
-    // cleaned up regardless, so subsequent tests can re-create their own `tbl`.
-    withTempDir { tempDir =>
-      withTable("tbl_geo_default_create") {
-        val ex = intercept[AnalysisException] {
-          sql(s"CREATE TABLE tbl_geo_default_create(g GEOMETRY($DefaultSrid) DEFAULT NULL) " +
-            s"USING delta LOCATION '${tempDir.getAbsolutePath}/t' " +
-            "TBLPROPERTIES('delta.feature.allowColumnDefaults' = 'supported')")
-        }
-        assert(ex.getCondition == "DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES",
-          s"Expected DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES, got: ${ex.getCondition}")
-      }
-    }
-  }
-
-  test("Column default values are rejected on geometry columns via ALTER COLUMN SET DEFAULT") {
-    // `ALTER TABLE ... ALTER COLUMN ... SET DEFAULT ...` goes through `ChangeColumn`, which is
-    // also matched by `checkColumnDefaults`.
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(g GEOMETRY($DefaultSrid)) USING delta " +
-        "TBLPROPERTIES('delta.feature.allowColumnDefaults' = 'supported')")
-      val ex = intercept[AnalysisException] {
-        sql("ALTER TABLE tbl ALTER COLUMN g SET DEFAULT NULL")
-      }
-      assert(ex.getCondition == "DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES",
-        s"Expected DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES, got: ${ex.getCondition}")
-    }
-  }
-
-  test("Column default on a STRUCT containing a geo field is rejected at CREATE TABLE") {
-    // `checkColumnDefaults` recurses through the column's data type via
-    // `typeExistsRecursively`, so a DEFAULT on a top-level struct column whose nested
-    // field is geo must also be rejected. This catches the regression where the default
-    // check stops at the top-level field type and ignores nested geo.
-    withTempDir { tempDir =>
-      withTable("tbl_geo_default_nested") {
-        val ex = intercept[AnalysisException] {
-          sql(s"CREATE TABLE tbl_geo_default_nested(" +
-            s"s STRUCT<inner: GEOMETRY($DefaultSrid)> DEFAULT NULL) " +
-            s"USING delta LOCATION '${tempDir.getAbsolutePath}/t' " +
-            "TBLPROPERTIES('delta.feature.allowColumnDefaults' = 'supported')")
-        }
-        assert(ex.getCondition == "DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES",
-          s"Expected DELTA_OPERATION_NOT_SUPPORTED_FOR_DATATYPES, got: ${ex.getCondition}")
-      }
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Universal Format (Hudi) compatibility
-  //
-  // `UniversalFormat.enforceHudiDependencies` rejects any schema containing a geo type
-  // when UniForm (Hudi) is enabled, throwing `DELTA_UNIVERSAL_FORMAT_VIOLATION`. These
-  // tests cover that code path on both the create-time and alter-time entry points.
-  // ---------------------------------------------------------------------------
-
-  test("UniForm (Hudi) rejects creating a table with a geometry column") {
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta " +
-          "TBLPROPERTIES(" +
-          "'delta.universalFormat.enabledFormats' = 'hudi', " +
-          "'delta.enableDeletionVectors' = 'false'" +
-          ")")
-      }
-      assert(ex.getErrorClass == "DELTA_UNIVERSAL_FORMAT_VIOLATION",
-        s"Expected DELTA_UNIVERSAL_FORMAT_VIOLATION, got: ${ex.getErrorClass}")
-      assert(ex.getMessage.toLowerCase(java.util.Locale.ROOT).contains("geometry"),
-        s"Expected message to mention GeometryType, got: ${ex.getMessage}")
-    }
-  }
-
-  test("UniForm (Hudi) rejects creating a table with a geography column") {
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(id INT, g GEOGRAPHY($DefaultSrid)) USING delta " +
-          "TBLPROPERTIES(" +
-          "'delta.universalFormat.enabledFormats' = 'hudi', " +
-          "'delta.enableDeletionVectors' = 'false'" +
-          ")")
-      }
-      assert(ex.getErrorClass == "DELTA_UNIVERSAL_FORMAT_VIOLATION",
-        s"Expected DELTA_UNIVERSAL_FORMAT_VIOLATION, got: ${ex.getErrorClass}")
-      assert(ex.getMessage.toLowerCase(java.util.Locale.ROOT).contains("geography"),
-        s"Expected message to mention GeographyType, got: ${ex.getMessage}")
-    }
-  }
-
-  test("UniForm (Hudi) rejects geo nested inside a struct") {
-    // Nested coverage: `findAnyTypeRecursively` is what `enforceHudiDependencies` calls,
-    // so a geo column buried inside a struct must also be rejected by the Hudi block.
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(s STRUCT<inner: GEOMETRY($DefaultSrid)>) USING delta " +
-          "TBLPROPERTIES(" +
-          "'delta.universalFormat.enabledFormats' = 'hudi', " +
-          "'delta.enableDeletionVectors' = 'false'" +
-          ")")
-      }
-      assert(ex.getErrorClass == "DELTA_UNIVERSAL_FORMAT_VIOLATION",
-        s"Expected DELTA_UNIVERSAL_FORMAT_VIOLATION, got: ${ex.getErrorClass}")
-    }
-  }
 
   // ---------------------------------------------------------------------------
   // DML read/write on actual geo data (Spark 4.2+ Parquet writer)
@@ -854,70 +254,6 @@ class DeltaGeoSuite extends QueryTest
     }
   }
 
-  // ---------------------------------------------------------------------------
-  // Schema evolution for geospatial columns through MERGE
-  // ---------------------------------------------------------------------------
-
-  test("Adding a new geo column via ALTER TABLE writes a Metadata action with the geo schema") {
-    withTable("tbl") {
-      sql("CREATE TABLE tbl(id INT, s STRING) USING delta")
-      sql(s"ALTER TABLE tbl ADD COLUMN g GEOGRAPHY($DefaultSrid)")
-      val deltaLog = DeltaLog.forTable(spark, TableIdentifier("tbl"))
-      val schema = deltaLog.update().metadata.schema
-      assert(schema("g").dataType.isInstanceOf[GeographyType])
-    }
-  }
-
-  test("SchemaMergingUtils.mergeSchemas rejects geo columns with different SRIDs " +
-    "(blocks SRID change / MERGE schema evolution with mismatched SRIDs)") {
-    // Two schemas that agree on column name and broad type (geometry) but differ in SRID. The
-    // geo guard in `SchemaMergingUtils.typeForImplicitCast` returns None for any geo-to-geo
-    // mismatch, so the merge falls through and the field-merge wrapper rethrows as
-    // `DELTA_FAILED_TO_MERGE_FIELDS` (whose cause is `DELTA_MERGE_INCOMPATIBLE_DATATYPE`). This
-    // is the same path that fires on `ALTER TABLE ... ALTER COLUMN g TYPE GEOMETRY(<other srid>)`
-    // and on `MERGE WITH SCHEMA EVOLUTION` when source/target SRIDs differ.
-    val tableSchema = new StructType().add("g", GeometryType(DefaultSrid))
-    val dataSchema = new StructType().add("g", GeometryType(0))
-    val ex = intercept[DeltaAnalysisException] {
-      SchemaMergingUtils.mergeSchemas(
-        tableSchema,
-        dataSchema,
-        allowImplicitConversions = true)
-    }
-    assert(ex.getErrorClass == "DELTA_FAILED_TO_MERGE_FIELDS",
-      s"Expected DELTA_FAILED_TO_MERGE_FIELDS, got: ${ex.getErrorClass}")
-  }
-
-  test("SchemaMergingUtils.mergeSchemas rejects geo->non-geo and non-geo->geo merges") {
-    // The geo guard short-circuits whenever EITHER side is geo, so swapping a non-geo column
-    // for a geo one (or vice versa) also fails to merge with `DELTA_FAILED_TO_MERGE_FIELDS`.
-    val geoToString = new StructType().add("g", GeometryType(DefaultSrid))
-    val stringToGeo = new StructType().add("g", StringType)
-    val ex1 = intercept[DeltaAnalysisException] {
-      SchemaMergingUtils.mergeSchemas(geoToString, stringToGeo, allowImplicitConversions = true)
-    }
-    assert(ex1.getErrorClass == "DELTA_FAILED_TO_MERGE_FIELDS")
-    val ex2 = intercept[DeltaAnalysisException] {
-      SchemaMergingUtils.mergeSchemas(stringToGeo, geoToString, allowImplicitConversions = true)
-    }
-    assert(ex2.getErrorClass == "DELTA_FAILED_TO_MERGE_FIELDS")
-  }
-
-  test("ALTER TABLE ALTER COLUMN ... TYPE between geo SRIDs is rejected") {
-    // `ALTER TABLE ... ALTER COLUMN g TYPE GEOMETRY(<other srid>)` routes through
-    // `AlterTableChangeColumnDeltaCommand`, which goes through schema merge. The geo guard
-    // above blocks the change. No data needs to be written, so this works even on OSS Spark
-    // 4.1 (whose Parquet writer cannot encode geo values).
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(g GEOMETRY($DefaultSrid)) USING delta")
-      val ex = intercept[AnalysisException] {
-        sql("ALTER TABLE tbl ALTER COLUMN g TYPE GEOMETRY(0)")
-      }
-      assert(ex.getMessage.toUpperCase(java.util.Locale.ROOT).contains("GEOMETRY"),
-        s"Expected the rejection message to mention GEOMETRY, got: ${ex.getMessage}")
-    }
-  }
-
   test("MERGE WITH SCHEMA EVOLUTION adds a new geo column from source to target") {
     // Real MERGE end-to-end. The source carries a geo column that the target does not yet
     // have; schema evolution adds the column and the matched/non-matched data is written
@@ -1001,30 +337,8 @@ class DeltaGeoSuite extends QueryTest
   }
 
   // ---------------------------------------------------------------------------
-  // Change Data Feed support for geospatial columns
+  // Change Data Feed lifecycle on geo data (Spark 4.2+ Parquet writer)
   // ---------------------------------------------------------------------------
-
-  test("Enabling CDF on a table with a geo column does not throw at table creation") {
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta " +
-        "TBLPROPERTIES('delta.enableChangeDataFeed' = 'true')")
-      val protocol = getProtocolForTable("tbl")
-      // GeoSpatial feature must be supported alongside CDF.
-      assert(protocol.isFeatureSupported(GeoSpatialTableFeature))
-    }
-  }
-
-  test("CDF schema contains the geo column with original type and SRID") {
-    withTable("tbl") {
-      sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta " +
-        "TBLPROPERTIES('delta.enableChangeDataFeed' = 'true')")
-      val deltaLog = DeltaLog.forTable(spark, TableIdentifier("tbl"))
-      val schema = deltaLog.update().metadata.schema
-      val geoField = schema("g")
-      assert(geoField.dataType.isInstanceOf[GeometryType])
-      assert(geoField.dataType.asInstanceOf[GeometryType].srid == DefaultSrid)
-    }
-  }
 
   test("CDF lifecycle: INSERT/UPDATE/DELETE on geo column shows up in table_changes") {
     // Full CDF lifecycle test (Spark 4.2-only; needs the Parquet writer to support geo):
@@ -1114,57 +428,6 @@ class DeltaGeoSuite extends QueryTest
         (1, Some(PointZeroZeroWkb), "insert"),
         (1, Some(PointTwoTwoWkb), "update_postimage"),
         (1, Some(PointZeroZeroWkb), "update_preimage")))
-    }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Iceberg / UniForm / Hudi compatibility for geospatial columns
-  // ---------------------------------------------------------------------------
-
-  test("IcebergCompatV1 is rejected on a table with a geometry column") {
-    // `CheckGeoSpatialTableFeatureDisabled` runs as part of `IcebergCompatV1.checks` and rejects
-    // the metadata before any conversion happens. V1 has no `CheckTypeInV2AllowList`, so this
-    // check is the only guard for geo columns under V1.
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta " +
-          "TBLPROPERTIES('delta.enableIcebergCompatV1' = 'true', " +
-          "'delta.columnMapping.mode' = 'name')")
-      }
-      assert(ex.getErrorClass == "DELTA_ICEBERG_COMPAT_VIOLATION.UNSUPPORTED_DATA_TYPE",
-        s"Unexpected error class: ${ex.getErrorClass}")
-    }
-  }
-
-  test("IcebergCompatV2 is rejected on a table with a geography column") {
-    // Same check as above, registered in `IcebergCompatV2.checks`. V2 also has
-    // `CheckTypeInV2AllowList` which would reject geo with a more generic error, but the
-    // dedicated geo check is ordered to fire first and produces the same error class.
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(id INT, g GEOGRAPHY($DefaultSrid)) USING delta " +
-          "TBLPROPERTIES('delta.enableIcebergCompatV2' = 'true', " +
-          "'delta.columnMapping.mode' = 'name')")
-      }
-      assert(ex.getErrorClass == "DELTA_ICEBERG_COMPAT_VIOLATION.UNSUPPORTED_DATA_TYPE",
-        s"Unexpected error class: ${ex.getErrorClass}")
-    }
-  }
-
-  test("Hudi UniForm on a table with a geometry column throws DELTA_UNIVERSAL_FORMAT_VIOLATION") {
-    // Covers the geo branch added to `UniversalFormat.enforceHudiDependencies`. The geo type
-    // is not representable in Hudi's schema, so we reject at commit time.
-    withTable("tbl") {
-      val ex = intercept[DeltaUnsupportedOperationException] {
-        sql(s"CREATE TABLE tbl(id INT, g GEOMETRY($DefaultSrid)) USING delta " +
-          "TBLPROPERTIES('delta.universalFormat.enabledFormats' = 'hudi', " +
-          "'delta.enableDeletionVectors' = 'false', " +
-          "'delta.columnMapping.mode' = 'name')")
-      }
-      assert(ex.getErrorClass == "DELTA_UNIVERSAL_FORMAT_VIOLATION",
-        s"Unexpected error class: ${ex.getErrorClass}")
-      assert(ex.getMessage.contains("hudi"),
-        s"Expected message to mention hudi, got: ${ex.getMessage}")
     }
   }
 }
